backend/requirements.txt
pyyaml
crewai[tools]
fastapi[standard]
boto3
pynamodb
bcrypt
tiktoken
pymupdf
PyJWT